+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix Arb Approval [#210](https://github.com/asgardex/asgardex-desktop/pull/210)
 - Remove BNB Bep2Chain & Client [#199](https://github.com/asgardex/asgardex-desktop/pull/200)
+- Action button removed. [#194] (https://github.com/asgardex/asgardex-desktop/pull/223)
 
 # 1.21.6 (2024-04-24)